Rename symlinked package.json file to avoid confusing installed packages
